Portfolio Desing
Portfolio Website using HTML & Bootstrap

This portfolio website is designed to showcase my skills, projects, and experience in web development. Built using HTML and Bootstrap, it offers a clean, responsive layout that adapts seamlessly to different devices. Key sections include an "About Me" page, a showcase of my completed projects with live demos, a blog for sharing insights, and a contact form for potential clients or collaborators. The design emphasizes simplicity and professionalism, utilizing Bootstrap's built-in grid system, typography, and components for smooth navigation and a user-friendly interface.